# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: GTW IOT API
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
...
securityDefinitions:
  UserSecurity:
    type: apiKey
    in: header
    name: x-api-key
security:
  - UserSecurity: []
paths:
...
  /data/getSingleData:
    x-swagger-router-controller: dataManage
    get:
      tags:
      - data
      summary: get a single data 
      description:  get a single data by device id and timestamp
      # used as the method name of the controller
      operationId: getSingleData
      parameters:
      - name: DeviceID
        in: query
        description: Device object thats need to be deleted
        required: true
        type: string
      - name: TimeStamp
        in: query
        required: true
        type: integer
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/DataModel"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /data/getSingleCalculatedData:
    x-swagger-router-controller: dataManage
    get:
      tags:
      - data
      summary: get a single data 
      description:  get a single data by Asset ID and time stamp
      # used as the method name of the controller
      operationId: getSingleCalculatedData
      parameters:
      - name: AssetID
        in: query
        description: Device object thats need to be deleted
        required: true
        type: string
      - name: TimeStamp
        in: query
        required: true
        type: integer
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/CalculatedDataModel"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /data/getMultipleCalculatedData:
    x-swagger-router-controller: dataManage
    get:
      tags:
      - data
      summary: get multiple calculated data
      description:  get multiple calculated data
      # used as the method name of the controller
      operationId: getMultipleCalculatedData
      parameters:
      - name: AssetID
        in: query
        description: Asset ID
        required: true
        type: string
      - name: StartTimeStamp
        in: query
        required: true
        type: integer
      - name: EndTimeStamp
        in: query
        required: true
        type: integer
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/CalculatedDataModel"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /data/getMultipleCalculatedDataWithParameter:
    x-swagger-router-controller: dataManage
    get:
      tags:
      - data
      summary: get multiple calculated data
      description:  get multiple calculated data
      # used as the method name of the controller
      operationId: getMultipleCalculatedDataWithParameter
      parameters:
      - name: AssetID
        in: query
        description: Asset ID
        required: true
        type: string
      - name: StartTimeStamp
        in: query
        required: true
        type: integer
      - name: EndTimeStamp
        in: query
        required: true
        type: integer
      - name: ParameterName
        in: query
        required: true
        type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/CalculatedDataModel"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /data/getMultipleData:
    x-swagger-router-controller: dataManage
    get:
      tags:
      - data
      summary: get multiple data 
      description:  get multiple data by device id and timestamp
      # used as the method name of the controller
      operationId: getMultipleData
      parameters:
      - name: DeviceID
        in: query
        description: Device object thats need to be deleted
        required: true
        type: string
      - name: StartTimeStamp
        in: query
        required: true
        type: integer
      - name: EndTimeStamp
        in: query
        required: true
        type: integer
      responses:
        200:
          description: Success
          schema:
            type: object
            properties:
              Items:
                type: array
                items:
                  $ref: "#/definitions/DataModel"
              Count:
                type: integer
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  CalculatedDataModel:
    type: object
    properties:
      AssetID:
        type: string
      EpochTimeStamp:
        type: integer
  DataModel:
    type: object
    properties:
      DeviceID:
        type: string
      EpochTimeStamp:
        type: integer
      Data:
        type: number
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string